+++ tests/dummy/app/routes/ember-flexberry-dummy-comment-vote-edit.js
@@ -1,0 +1,22 @@
+import EditFormRoute from 'ember-flexberry/routes/edit-form';
+import EditFormRouteOperationsIndicationMixin from '../mixins/edit-form-route-operations-indication';
+
+export default EditFormRoute.extend(EditFormRouteOperationsIndicationMixin, {
+  /**
+    Name of model projection to be used as record's properties limitation.
+
+    @property modelProjection
+    @type String
+    @default 'CommentVoteE'
+   */
+  modelProjection: 'CommentVoteE',
+
+  /**
+    Name of model to be used as form's record type.
+
+    @property modelName
+    @type String
+    @default 'ember-flexberry-dummy-comment-vote'
+   */
+  modelName: 'ember-flexberry-dummy-comment-vote'
+});